backend/requirements-test.txt
pytest
pytest-django
pytest-cov
pytest-html
djangorestframework
